+++ src/utils/request.ts
@@ -1,18 +1,35 @@
-import { getQuery } from 'ufo'
+import { getQuery as _getQuery } from 'ufo'
 import { createError } from '../error'
 import type { HTTPMethod } from '../types'
 import type { CompatibilityEvent } from '../event'
 
-export function useQuery (event: CompatibilityEvent) {
-  return getQuery(event.req.url || '')
+export function getQuery (event: CompatibilityEvent) {
+  return _getQuery(event.req.url || '')
 }
 
-export function useMethod (event: CompatibilityEvent, defaultMethod: HTTPMethod = 'GET'): HTTPMethod {
+/** @deprecated Use `h3.getQuery` */
+export const useQuery = getQuery
+
+export function getRouterParams (event: CompatibilityEvent): CompatibilityEvent['context'] {
+  // Fallback object needs to be returned in case router is not used (#149)
+  return event.context.params || {}
+}
+
+export function getRouterParam (event: CompatibilityEvent, name: string): CompatibilityEvent['context'][string] {
+  const params = getRouterParams(event)
+
+  return params[name]
+}
+
+export function getMethod (event: CompatibilityEvent, defaultMethod: HTTPMethod = 'GET'): HTTPMethod {
   return (event.req.method || defaultMethod).toUpperCase() as HTTPMethod
 }
 
+/** @deprecated Use `h3.getMethod` */
+export const useMethod = getMethod
+
 export function isMethod (event: CompatibilityEvent, expected: HTTPMethod | HTTPMethod[], allowHead?: boolean) {
-  const method = useMethod(event)
+  const method = getMethod(event)
 
   if (allowHead && method === 'HEAD') {
     return true
@@ -37,3 +54,18 @@ export function assertMethod (event: CompatibilityEvent, expected: HTTPMethod | 
     })
   }
 }
+
+export function getRequestHeaders (event: CompatibilityEvent): CompatibilityEvent['req']['headers'] {
+  return event.req.headers
+}
+
+export const getHeaders = getRequestHeaders
+
+export function getRequestHeader (event: CompatibilityEvent, name: string): CompatibilityEvent['req']['headers'][string] {
+  const headers = getRequestHeaders(event)
+  const value = headers[name.toLowerCase()]
+
+  return value
+}
+
+export const getHeader = getRequestHeader