+++ src/utils/response.ts
@@ -1,10 +1,11 @@
+import type { OutgoingMessage } from 'http'
 import { createError } from '../error'
 import type { CompatibilityEvent } from '../event'
 import { MIMES } from './consts'
 
 const defer = typeof setImmediate !== 'undefined' ? setImmediate : (fn: Function) => fn()
 
-export function send (event: CompatibilityEvent, data: any, type?: string): Promise<void> {
+export function send (event: CompatibilityEvent, data?: any, type?: string): Promise<void> {
   if (type) {
     defaultContentType(event, type)
   }
@@ -23,12 +24,45 @@ export function defaultContentType (event: CompatibilityEvent, type?: string) {
 }
 
 export function sendRedirect (event: CompatibilityEvent, location: string, code = 302) {
+  const encodedLoc = encodeURI(location)
   event.res.statusCode = code
-  event.res.setHeader('Location', location)
-  return send(event, 'Redirecting to ' + location, MIMES.html)
+  event.res.setHeader('Location', encodedLoc)
+  // Minimal html document that redirects on client side
+  const html = `<!DOCTYPE html>
+<html>
+  <head><meta http-equiv="refresh" content="0; url=${encodedLoc}"></head>
+  <body>Redirecting to <a href=${JSON.stringify(encodedLoc)}>${encodedLoc}</a></body>
+</html>`
+  return send(event, html, MIMES.html)
 }
 
-export function appendHeader (event: CompatibilityEvent, name: string, value: string): void {
+export function getResponseHeaders (event: CompatibilityEvent): ReturnType<CompatibilityEvent['res']['getHeaders']> {
+  return event.res.getHeaders()
+}
+
+export function getResponseHeader (event: CompatibilityEvent, name: string): ReturnType<CompatibilityEvent['res']['getHeader']> {
+  return event.res.getHeader(name)
+}
+
+export function setResponseHeaders (event: CompatibilityEvent, headers: Record<string, Parameters<OutgoingMessage['setHeader']>[1]>): void {
+  Object.entries(headers).forEach(([name, value]) => event.res.setHeader(name, value))
+}
+
+export const setHeaders = setResponseHeaders
+
+export function setResponseHeader (event: CompatibilityEvent, name: string, value: Parameters<OutgoingMessage['setHeader']>[1]): void {
+  event.res.setHeader(name, value)
+}
+
+export const setHeader = setResponseHeader
+
+export function appendResponseHeaders (event: CompatibilityEvent, headers: Record<string, string>): void {
+  Object.entries(headers).forEach(([name, value]) => appendResponseHeader(event, name, value))
+}
+
+export const appendHeaders = appendResponseHeaders
+
+export function appendResponseHeader (event: CompatibilityEvent, name: string, value: string): void {
   let current = event.res.getHeader(name)
 
   if (!current) {
@@ -42,6 +76,8 @@ export function appendHeader (event: CompatibilityEvent, name: string, value: st
 
   event.res.setHeader(name, current.concat(value))
 }
+
+export const appendHeader = appendResponseHeader
 
 export function isStream (data: any) {
   return data && typeof data === 'object' && typeof data.pipe === 'function' && typeof data.on === 'function'