+++ test/app.test.ts
@@ -19,6 +19,23 @@ describe('app', () => {
     expect(res.body).toEqual({ url: '/' })
   })
 
+  it('can return a 204 response', async () => {
+    app.use('/api', () => null)
+    const res = await request.get('/api')
+
+    expect(res.statusCode).toBe(204)
+    expect(res.text).toEqual('')
+    expect(res.ok).toBeTruthy()
+  })
+
+  it('can return primitive values', async () => {
+    const values = [true, false, 42, 0, 1]
+    for (const value of values) {
+      app.use(`/${value}`, () => value)
+      expect(await request.get(`/${value}`).then(r => r.body)).toEqual(value)
+    }
+  })
+
   it('can return Buffer directly', async () => {
     app.use(() => Buffer.from('<h1>Hello world!</h1>', 'utf8'))
     const res = await request.get('/')